Make the descriptions of the CLI options consistent
The help texts for the CLI flags at main.go mostly matched those in usage.txt,
but there were a few slight discrepancies. This commit fixes them.

The exception is the description of the -h/-help option,
which is an entirely different sentence from that in usage.txt.
The only change made to it was changing the verb tense to match the others.